docs/requirements.txt
sphinx-rtd-theme==1.0.0
sphinx_autodoc_typehints==1.19.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
autodocsumm==0.2.9
sphinx_autodoc_defaultargs==0.1.2
fractal-tasks-core==0.3.4
